+++ e2e/source/test/scala/org/knora/webapi/SegmentE2EZSpec.scala
@@ -3,14 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package org.knora.webapi.e2ez
-
-import zio._
-import zio.json._
-import zio.test._
-
-import org.knora.webapi.E2EZSpec
-import org.knora.webapi.messages.store.triplestoremessages.RdfDataObject
+package org.knora.webapi
 
 object SegmentE2EZSpec extends E2EZSpec {
